+++ src/util/asserters.ts
@@ -1,3 +1,4 @@
+import {ServerError} from '@sanity/client'
 import {type InputProps, isObjectInputProps, type PreviewLayoutKey, type PreviewProps} from 'sanity'
 
 import type {MuxInputPreviewProps, MuxInputProps} from './types'
@@ -19,4 +20,17 @@ export function isValidUrl(url: string): boolean {
   } catch {
     return false
   }
+}
+
+/**
+ * We consider a server error one with status code 5XX.
+ * Used mainly to handle unknown Proxy issues.
+ */
+export function isServerError(error: Error): error is ServerError {
+  return (
+    'statusCode' in error &&
+    typeof error.statusCode === 'number' &&
+    500 <= error.statusCode &&
+    error.statusCode <= 600
+  )
 }